refactor: create a share-ui component AddNewBar
