+++ packages/shared-ui/src/index.ts
@@ -10,6 +10,7 @@ import DatePickerContainer from './components/DatePicker';
 import BorderlessDatePickerContainer from "./components/DatePicker/Borderless";
 import ScrollbarContainer from "./components/Scrollbar";
 import FormGroupContainer from "./components/FormGroup";
+import AddNewBarContainer from "./components/AddNewBar";
 
 export { messageInfo, messageError, messageSuccess, messageWarning } from './components/Message';
 
@@ -23,6 +24,7 @@ export const DatePicker = DatePickerContainer;
 export const DatePickerBorderless = BorderlessDatePickerContainer;
 export const Scrollbar = ScrollbarContainer
 export const FormGroup = FormGroupContainer
+export const AddNewBar = AddNewBarContainer
 export * from './components/Confirmbox'
 
 export * from './components/Controls/ListControl/type';
