+++ packages/ui-app/app/[orgID]/project/[projectId]/settings/status/index.tsx
@@ -1,33 +1,20 @@
-import { KeyboardEvent, useRef } from 'react'
-import { AiOutlinePlus } from 'react-icons/ai'
-import { StatusDnDContainer } from './StatusDnDContainer'
+import { useStatus } from '@/hooks/status'
 import { DndProvider } from 'react-dnd'
 import { HTML5Backend } from 'react-dnd-html5-backend'
-import { useStatus } from '@/hooks/status'
+
+import { StatusDnDContainer } from './StatusDnDContainer'
+import { AddNewBar } from '@shared/ui'
 
 export const ProjectStatus = () => {
-  const inputAddRef = useRef<HTMLInputElement>(null)
   const { createNewStatus } = useStatus({})
-
-  const onPressEnter = (e: KeyboardEvent<HTMLDivElement>) => {
-    createNewStatus(e, inputAddRef)
-  }
 
   return (
     <div className="w-full">
-      <div className="setting-container border dark:border-gray-700">
+      <div className="border setting-container dark:border-gray-700">
         <DndProvider backend={HTML5Backend}>
           <StatusDnDContainer />
         </DndProvider>
-        <div className="relative flex items-center bg-gray-50 dark:bg-gray-800 rounded-b-lg">
-          <AiOutlinePlus className="absolute top-3.5 left-4 text-gray-500" />
-          <input
-            ref={inputAddRef}
-            className="bg-transparent w-full pl-12 text-gray-500 text-sm pr-8 py-3 outline-none"
-            placeholder="Hit `Enter` to add a new status"
-            onKeyDown={onPressEnter}
-          />
-        </div>
+        <AddNewBar createNewAction={createNewStatus} itemName="status" />
       </div>
     </div>
   )
